Refactor file stucture. Add dependency.
