+++ src/Components/Table/DataStorageClasses/TableAttributesInfo.tsx
