+++ src/Components/Table/LoadingAnimation/BasicLoadingIcon.tsx
