+++ src/Components/Table/enums/RestrictionType.tsx
